+++ src/vm/interface/scene2d.cpp
@@ -20,9 +20,11 @@ void UpdateBackground2D() {
     int bufId = ScrWork[SW_BG1SURF + bgId];
 
     if (Profile::UseScreenCapEffects) {
-      if (GetScriptBufferId(bgId) == ScrWork[SW_EFF_CAP_BUF]) {
+      if (GetScriptBufferId(static_cast<int>(bgId)) ==
+          ScrWork[SW_EFF_CAP_BUF]) {
         Backgrounds2D[bufId] = &Screencaptures[0];
-      } else if (GetScriptBufferId(bgId) == ScrWork[SW_EFF_CAP_BUF2]) {
+      } else if (GetScriptBufferId(static_cast<int>(bgId)) ==
+                 ScrWork[SW_EFF_CAP_BUF2]) {
         Backgrounds2D[bufId] = &Screencaptures[1];
       } else if (Backgrounds2D[bufId]->BgSprite.Sheet.IsScreenCap) {
         Backgrounds2D[bufId] = &Backgrounds[bufId];
